synaptics: Immediately complete verification if finger is out already
When quickly scanning a finger with the synaptic driver, it may wait forever
for finger removal even if this has already happened.

In fact we don't take care of the finger status when reporting the
verification.

To avoid this, add a function that reports the verification state and that
delays its completion to the finger removal if the finger is on sensor,
otherwise it just does it.

Fixes #228