vfs0090: Inherit from more powerful FpDevice
Manually implement some imaging capabilities, by making the sensor a generic
one (as it actually is) with capture capabilities.

This implies some minimal code duplication, but also simplifies the behavior
while at the same time, we've now more control on the device and we can
implement with more fine grained control, an example is already in this
commit where we support the early-match mechanism, and so we notify early
whether we've a match or not, while we take some more time to deactivate
the device (i.e. reset the device back to sleepy state).

We now also properly handle cancellations during open and early phases.