pa_primex: Properly initialize set to invalid data the match index array
We were initializing the matched_index only in the case the device was
not matching, but in case we had a match, we were resetting the array
with an index array that was always initialized to zero.

This was particularly wrong when we were trying to match the finger with
device id set to 0, as in such case any other finger that was matching
was accepted.