device: Return an error if trying to verify / identify using non-device stored prints
In case we are verifying using a print template that is not stored
anymore into the device storage we may want to delete the print
reference locally, however not all the drivers may support this
natively.

Said this, if a device has storage and does not support checking for
templates data presence on storage during such stage we may want to
do it once the verify / identify operation finished with a match failure
(not to waste time before and assuming that we'd already notified using
early reporting).

So in such case, go through the list of device saved prints and if that
is not containing the required data, return an error instead of just
simple match failure.

Adapt tests.