fp-device: Keep storage device prints cached if no action is performed
As per previous commit we may end up requesting a device to list its
storage contents multiple times when the device is active for
verification.

Given that during the time we keep a device opened we can be sure that
nothing else will change its storage, we can keep a prints cache around
and change or delete it only when doing destructive operations (such as
enroll, delete and closing the device).

Add tests