Optionally use native generators
fprint-list-udev-hwdb and fprint-list-udev-rules are run
during the build to generate autosuspend.hwdb and 70-*.rules,
respectively.

Since they are not marked as "native: true", a cross-compiled
version is not possible.

Since these binaries are linked with the libfprint_drivers target,
marking them as native binaries would also need libfprint_drivers
to be duplicated as a native version and possibly other library
targets, too.

Instead, make it the responsibility of the cross-compiler
framework to build the native variant separately and allow
the external binaries to be passed in.

Signed-off-by: Zoltán Böszörményi <zboszor@gmail.com>